Avanzo #25. Saco calculo del costo ya que no es especifico del test
